pyproject.toml
[project]
name = "ds-bookstore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpc-stubs>=1.53.0.6",
    "grpcio-tools>=1.78.0",
    "mypy-protobuf>=5.0.0",
    "python-levenshtein>=0.27.3",
    "requests>=2.32.5",
    "ruff>=0.15.10",
]
